fix(scripts): bound fixture teardown with the shared Electron close
withFixtureWindow tore down with a bare app.close(), which has no
deadline — one wedged launch turned a capture run (or the CI alignment
audit, which has no job timeout) into an infinite hang, and the temp
userData dir leaked because rm was sequenced after the close.

The bounded close the E2E suite already owned moves to
scripts/electron-lifecycle.mjs — the same shared-home precedent as
fixture-env.mjs, so a bare-node harness and the Playwright suite stop
splitting one launch concern across two directories. Renderer evaluate
gets a deadline too; teardown then reclaims the process either way.
